requirements.txt
Django
djangorestframework
django-filter
psycopg2
drf_spectacular